Move BufferEditor actions into a separate Action structs
Instead of handling events directly, we use Actions now. This looks
like an overkill but won't when there is an undo/redo system. Undo/
redo systems require an interpreter over actions and state-awareness
of actions, that is, actions should be able to capture/negotiate with
state to be able to undo/redo properly. At least I think so...

Also a bunch of renames because I can't live a day without renaming
things...