Move protocol, rule, signatures to a different file
Also remove message id, sender info because they are unused and won't
be used. Eases the load a bit on the system (hopefully...)

This means messages are no longer unique. Also, the receiver cell
does not know nothing about itself nor the sender because cells are
not actually the "subjects" -- protocols and rules are, and they
don't know anything about the cell that is adhering to them/expressing
them. A bit similar to nature (as far as I am aware!!) where genes don't
really care who expresses them.